ICT2-44 코드리뷰 및 수정, exception 처리 추가
